adding kyverno test command with git support
